refactor: Kid, Parent의 구분이 필요없는 페이지에 대해서는 prefix 제거
