+++ src/pages/Home/Walking.tsx
